% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTTESim.R
\name{checkTTESim}
\alias{checkTTESim}
\title{Check simulations data}
\usage{
checkTTESim(
  sdata,
  odata = NULL,
  rtte = FALSE,
  idVar = "ID",
  dvVar = "DV",
  timeVar = "TIME",
  rtteVar = "RTTE",
  surVar = "SUR",
  iCountVar = "ICOUNT",
  iterVar = "ITER",
  randVar = "RAND"
)
}
\arguments{
\item{sdata}{a data.frame, simulated data. Ideally the output from NONMEM without modifications.}

\item{odata}{a data.frame, observed data. Only event records. If missing, checks that compare odata and sdata will be omitted.}

\item{rtte}{a logical, indicating if sdata and odata are repeated time to event data (TRUE).}

\item{idVar, dvVar, timeVar, rtteVar, surVar, iCountVar, iterVar, randVar}{Character string specifying the name of the columns. Defaulted to the output of a NONMEM code generated with \code{createTTESim()}.}
}
\value{
Used for its side effects: returns warnings if one or multiple checks fail.
}
\description{
This function checks several assumptions expected from a TTE/RTTE data set simulated using NONMEM.
Ideally, it should be called directly after the NONMEM table is imported in R, before any further modifications.
If a check fails, a warning will be returned.
}
\details{
Following assumptions are checked:
\itemize{
\item The number of ID is the same within each iteration
\item DV are only 0 or 1
\item RTTE is computed correctly (given ITER/ID)
\item ICOUNT is computed correctly (given ITER/ID)
\item IDs are the same in odata and sdata
\item Dimension of the sdata is a multiple of the odata (multiple records per ID allowed for RTTE data)
}
}
\examples{
# Dummy legit data
ex_odata <- data.frame(ID = c(1, 4, 7), DV = c(0, 1, 1), TIME = c(100, 150, 600))
ex_sdata_ok <- data.frame(
  ID = c(1, 4, 7, 1, 4, 7, 1, 4, 7),
  DV = c(1, 0, 0, 0, 1, 0, 0, 0, 1),
  TIME = c(50, 400, 700, 100, 200, 700, 100, 400, 300),
  RTTE = c(1, 1, 1, 1, 1, 1, 1, 1, 1),
  SUR = runif(9), # not used in checks
  ICOUNT = c(1, 2, 3, 4, 5, 6, 7, 8, 9),
  ITER = c(1,  1, 1, 2, 2, 2, 3, 3, 3),
  RAND = runif(9) # not used in checks
  )

# If all ok, remains silent
checkTTESim(sdata = ex_sdata_ok)

# Additional checks if odata is provided
checkTTESim(sdata = ex_sdata_ok, odata = ex_odata)

# Will warn if error:
ex_sdata_wrong <- ex_sdata_ok[-5,] # remove one record
\dontrun{
  checkTTESim(sdata = ex_sdata_wrong, odata = ex_odata)
}
}
