% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTTEStart.R
\name{createTTEStart}
\alias{createTTEStart}
\title{Create NONMEM control stream for Time-To-Event (TTE) models}
\usage{
createTTEStart(
  outFile = NULL,
  rtte = FALSE,
  dataFile = NULL,
  replaceAsId = NULL,
  replaceAsTime = NULL,
  distribution = "exponential",
  includeTable = TRUE
)
}
\arguments{
\item{outFile}{Character string specifying the filename (and path) for the
NONMEM estimation file to be written to disk. If \code{NULL}, no file is
written, and the model code is returned as a character vector.}

\item{rtte}{Logical, indicating whether to create a model for repeated
time-to-event simulations (\code{TRUE}) or standard time-to-event simulations
(\code{FALSE}, default).}

\item{dataFile}{Character string specifying the path to a data file to be
used for the \verb{$INPUT} and \verb{$DATA} sections of the NONMEM control stream.
If provided, the function will attempt to read the headers from this file
to create the \verb{$INPUT} statement. If \code{NULL} (default), a dummy \verb{$DATA}
section is created, and a basic \verb{$INPUT ID TIME DV EVID} is used.}

\item{replaceAsId}{Character string. If a data file is provided via
\code{dataFile}, this argument allows you to specify a string in the data file
header that should be replaced with "ID" in the \verb{$INPUT} statement.
Default is \code{NULL} (no replacement).}

\item{replaceAsTime}{Character string.  If a data file is provided via
\code{dataFile}, this argument allows you to specify a string in the data file
header that should be replaced with "TIME" in the \verb{$INPUT} statement.
Default is \code{NULL} (no replacement).}

\item{distribution}{Character string specifying the survival distribution
to be used. Options are: \code{"exponential"} (default), \code{"weibull"},
\code{"weibull_exp"}, \code{"gompertz"}, and \code{"no"} (for a user-defined hazard
function, where "PAR=THETA(1)*EXP(ETA(1))" is used).}

\item{includeTable}{Logical, indicating whether to include a \verb{$TABLE}
section in the NONMEM control stream. Only relevant if \code{dataFile} is not
\code{NULL}. Default is \code{TRUE}.}
}
\value{
If \code{outFile} is specified, the function writes the NONMEM control
stream to the specified file and returns \code{NULL} invisibly. If \code{outFile}
is \code{NULL}, the function returns the NONMEM control stream as a character
vector.
}
\description{
This function generates a NONMEM control stream file for
time-to-event (TTE) or repeated time-to-event (RTTE) models. It supports
different survival distributions and allows for data input from a file.
}
\examples{
# Example TTE model with Weibull distribution and output to file
# File path to a data object (embedded in the package as an example)
mypath <- system.file("extdata", "tte_data1.dat", package = "PMXtte")
createTTEStart(
  outFile = "tte_weibull_auto.mod",
  dataFile = mypath,
  distribution = "weibull",
  rtte = FALSE
)

# Example RTTE model with Weibull distribution, data from CSV,
# replacing "TSFDY" with "TIME" in $INPUT
mypath <- system.file("extdata", "DAT-1a-PMX-RTTEWorkshop-PFPMX-1.csv", package = "PMXtte")
createTTEStart(
  outFile = "rtte_mod.mod",
  dataFile = mypath,
  distribution = "weibull",
  rtte = TRUE,
  replaceAsTime = "TSFDY"
)
}
