% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_rtte.R
\name{factor_rtte}
\alias{factor_rtte}
\title{Factor the RTTE variable}
\usage{
factor_rtte(x)
}
\arguments{
\item{x}{a vector of numeric}
}
\value{
a vector of factor
}
\description{
Factor the RTTE variable
}
\examples{
ex <- data.frame(
  ID = c(1,1,2,2,2),
  RTTE = c(1,2,1,2,3),
  DV = c(1,0,1,1,0),
  TIME = c(100, 200, 50, 150, 250)
)
dplyr::mutate(ex, RTTEF = factor_rtte(RTTE))
# Useful to facet Kaplan-Meier with RTTE data
ggKAP(
 data = dplyr::mutate(ex, RTTEF = factor_rtte(RTTE)),
 facet_var = "RTTEF"
)
}
