% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggKAP.R
\name{ggKAP}
\alias{ggKAP}
\title{Visualize Survival Data}
\usage{
ggKAP(
  data,
  sdata,
  time_var = "TIME",
  dv_var = "DV",
  iter_var = "ITER",
  col_var = NULL,
  facet_var = NULL,
  show_kpm = !missing(data),
  show_censor = !missing(data),
  show_risktable = !missing(data),
  show_se = !missing(data) && missing(sdata),
  show_pval = !missing(data) && missing(sdata) && !is.null(col_var) && (if
    (!is.null(facet_var)) {
     col_var != facet_var
 } else {
     TRUE
 }),
  show_median = !missing(data) && missing(sdata),
  show_sim = !missing(sdata),
  ci_level = if (show_se) 0.95 else 0.9,
  cuminc = FALSE,
  ci_alpha = 0.5,
  censor_shape = "|",
  censor_size = 4,
  median_linetype = 2,
  scale_x_break_by = NULL,
  scale_y_labels = scales::percent,
  label_x = "Time since first dose (weeks)",
  label_y_fig = paste0("Subjects with", if (!cuminc) {
     "out"
 }, " event (\%)"),
  label_col = col_var,
  label_y_risktab = label_col,
  legend_label_kpm = "Observed",
  legend_label_censor = "Censored",
  legend_label_se = paste0(scales::percent(ci_level), " CI"),
  legend_label_median = "Median Survival",
  legend_label_sim = paste0("Simulated ", scales::percent(ci_level), " CI"),
  scale_color_labels = waiver(),
  scale_color_values = unname(PMXColors_PMXcolors$default),
  facetting_args = list(),
  xlim = c(NA_real_, NA_real_),
  ylim = c(0, 1),
  title_risktable = "Number of subjects at risk",
  arrange = TRUE,
  rel_heights = list(legends = c(1, 1), figtable = c(8, 4), overall = c(1, 8))
)
}
\arguments{
\item{rel_heights}{}
}
\description{
Visualize Survival Data
}
\examples{

dat <- PMXtte::simplettedata
simdat <- PMXtte::simplettesimdata

ggKAP(dat)
ggKAP(dat, show_kpm = FALSE)
ggKAP(dat, show_censor = FALSE)
ggKAP(dat, show_se = FALSE)
ggKAP(dat, show_risktable = FALSE)
ggKAP(dat, show_median = FALSE)
try(ggKAP(dat, show_sim = TRUE), TRUE)


ggKAP(dat, col_var = "SEXF")
ggKAP(dat, col_var = "SEXF", show_pval = FALSE)
ggKAP(dat, col_var = "SEXF", scale_color_values = c("red", "yellow"))
ggKAP(dat, col_var = "SEXF", scale_color_values = c(Male = "yellow", Female = "red"))
ggKAP(dat, col_var = "SEXF", ci_alpha = .3)
ggKAP(dat, col_var = "SEXF", ci_alpha = .7)

ggKAP(dat, rel_heights = list(
  legends = c(1,1),
  figtable = c(8,2),
  overall = c(1,8)
))
ggKAP(dat, col_var = "SEXF", facet_var = "AUCQF", rel_heights = list(
  legends = c(1,1),
  figtable = c(8,4),
  overall = c(2,8)
))

ggKAP(dat, simdat, facet_var = "SEXF")
ggKAP(sdata = simdat, facet_var = "SEXF")

ggKAP(dat, col_var = "AUCQF")
ggKAP(dat, col_var = "AUCQF", label_col = expression(AUC['0-24h']))
dat$AUCQF2 <- factor(dat$AUCQF, labels = c("1^st", "2^nd~quartile", "The~3^rd~q"))
ggKAP(dat, col_var = "AUCQF2", label_col = expression(AUC['0-24h']), scale_color_labels = scales::label_parse())

ggKAP(dat, facet_var = "AUCQF2")
ggKAP(dat, facet_var = "AUCQF2", facetting_args = list(labeller = ggplot2::label_parsed, nrow = 2))
}
