% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggKAP.R
\name{ggKAParrange}
\alias{ggKAParrange}
\title{Arrange ggKAP() results on a single plot}
\usage{
ggKAParrange(x, legends, figtable, overall)
}
\arguments{
\item{x}{a named list, returned by \code{ggKAP(arrange = FALSE)}. See Details for the expected names.}

\item{legends, figtable, overall}{relative heights (1) between the legends (default is 1 and 1, same height), (2) between the Kaplan-Meier figure and the risk table (default is 2 and 1, i.e. figure twice higher than table) and (3) overall (default is 1 and 8, combined legends takes 1/9 of heights, combined figure and tables takes 8/9 of heights)}
}
\value{
a ggplot object
}
\description{
Arrange ggKAP() results on a single plot
}
\details{
Expected names are: \code{leg_col} (legends of the categories of the coloring variable),
\code{leg_kpm} (legends of the different objects on the Kaplan-Meier figure),
\code{fig} (ggplot object for the Kaplan-Meier figure),
and \code{risktab} (ggplot object for the risk table).
}
\examples{
library(ggplot2)
dat <- PMXtte::simplettedata

plotlist <- ggKAP(dat, color_var = "SEXF", arrange = FALSE)

# The Kaplan-Meier figure is in $fig
# The Risk table is in $risktab
# Custom these elements to your needs:
plotlist$fig <- plotlist$fig +
  theme(
    text = element_text(size = 20),
    axis.text.x = element_text(angle = 90, hjust = 1)
    )

# Re-arrange them with ggKAParrange:
ggKAParrange(plotlist)
ggKAParrange(plotlist, figtable = c(1,2))
}
