% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggKMMC.R
\name{ggKMMC}
\alias{ggKMMC}
\title{Kaplan-Meier Mean Covariate plots for TTE models}
\usage{
ggKMMC(
  odata,
  sdata = NULL,
  cov_col,
  cov_data = odata,
  time_col = "TIME",
  id_col = "ID",
  iter_col = "ITER",
  grouping_col = character(0),
  bins = seq(0, max(odata[[time_col]]), length.out = 50),
  fun = "mean",
  ylab = paste(stringr::str_to_title(fun), "of", cov_col),
  nocb = TRUE,
  output_data = FALSE,
  obsCol = "black",
  simFill = PMXtte:::PMXColors_pmx_palettes(name = "light"),
  ci = c(0.05, 0.95)
)
}
\arguments{
\item{odata}{observed data}

\item{sdata}{simulated data, the number of rows must be proportional to the number of rows in \code{odata}}

\item{cov_col}{a character, the covariate column to compute the rolling mean of. Must be present in \code{cov_data}.}

\item{cov_data}{covariate data. The default is \code{odata}, suitable for \emph{non time-varying covariates}. For time-varying covariates, a data set with multiple records per ID that inform the changes in covariate values, should be used, typically a "derived data set"}

\item{time_col}{a character, the time column. The default is \code{"TIME"}.  Must be present in \code{odata}, \code{sdata} and \code{cov_data}.}

\item{id_col}{a character, the subject identifier column. The default is \code{"ID"}.  Must be present in \code{odata}, \code{sdata} and \code{cov_data}.}

\item{iter_col}{a character, the simulation replicate identifier column. The default is \code{"ITER"}.  Must be present in\code{sdata}.}

\item{grouping_col}{a character, the grouping variables, useful to facet the plot. The default is \code{character(0)}, overall data. Must be present in \code{odata}, and \code{sdata}.}

\item{bins}{a numeric, vector of numeric values when the covariate mean should be calculated. Default is set to have 50 values whatever the input data.}

\item{fun}{a character, name of the function to summarize the covariate. Default is \code{"mean"}.}

\item{ylab}{a character, label of the y-axis.}

\item{nocb}{a logical, should covariate value be interpolated with the "next observation carried backward" rule (the default)? Otherwise "last observation carried forward" will be used}

\item{output_data}{a logical, should the function return the data that build the plot? If not, the default, a figure is returned}

\item{obsCol}{a character, the color of the "observed mean" line}

\item{simFill}{a character, the filling color of the "simulated CI" ribbon}

\item{ci}{a numeric of length 2, with values between 0 and 1, the width of the confidence interval. Default is a 90\% CI.}
}
\value{
By default, a ggplot object. If \code{output_data = TRUE}, a data.frame to build the plot.
}
\description{
Kaplan-Meier Mean Covariate plots for TTE models
}
\details{
\subsection{Time varying covariates}{

This function calculates the running mean of a covariate as a function of time.
This mean is calculated at each time point defined in \code{bins}, after exclusion of subjects who experienced an event or censoring.
For time-varying covariates, the covariate values are interpolated following the "next observation carried backward" rule (unless \code{nocb=FALSE}).
Thus, it is of importance to use the argument \code{cov_data} with a data set that is rich enough to adequately describe the variation of the covariate in time.
Otherwise, only the covariate value in \code{odata} will be read, which will lead to a biased figure.
}
}
\examples{
library(magrittr)
derData <- readr::read_csv(
  system.file('extdata/DAT-1c-RED-1a-PMX-WOWTTE-PFPMX-1.csv', package= 'PMXtte'),
  show_col_types = FALSE) \%>\%
  dplyr::rename(TIME = TSFDW, SEX = SEXN, DOSE = DOSEN, COSEASON = COSEAS
 )
odata <- dplyr::filter(derData, EVID == 0, TYPE == 2, STUDYIDN == 1) \%>\%
  filter_xth_event(1)
sdata <- readRDS(system.file('extdata/vpcdat.rds', package= 'PMXtte')) \%>\%
  dplyr::group_by(ITER) \%>\%
  filter_xth_event(1) \%>\%
  dplyr::ungroup() \%>\%
  dplyr::mutate(TIME = TIME / (24*7), SEX = SEX+1)

# Basic plot
ggKMMC(odata = odata, sdata = sdata, cov_col = "DOSE")

# Works with binary covariate
ggKMMC(odata = odata, sdata = sdata, cov_col = "SEX")

# For time-varying covariate: use data with records of time-varying values
# e.g. the "derived data set" which still has records with `EVID==2`
ggKMMC(
  odata = odata, sdata = sdata,
  cov_col = "COSEASON", cov_data = derData,
  bins = seq(0,52,1)
)

# Grouping variables
ggKMMC(odata = odata, sdata = sdata, cov_col = "DOSE", grouping_col = "SEX")


# Minimal customization
p <- ggKMMC(odata = odata, sdata = sdata, cov_col = "DOSE", obsCol = "blue", simFill = "yellow")
p

# Customize as a regular ggplot object
p +
 ggplot2::labs(y = "Mean dose (mg)", x = "Time since first dose (weeks)")

# Output the underlying data and make your own plot if you prefer
ggKMMC(odata = odata, sdata = sdata, cov_col = "DOSE", output_data = TRUE)

}
