% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggKMvpc.R
\name{ggKMvpc}
\alias{ggKMvpc}
\title{Title}
\usage{
ggKMvpc(
  odata,
  sdata,
  time = TIME,
  event = DV,
  iter = ITER,
  strat = NULL,
  ci = c(0.05, 0.95),
  cuminc = T,
  simCol = "blue",
  obsCol = "blue",
  posObs = NULL,
  show.censor = T,
  censor.shape = "|",
  censor.size = 4,
  palette = pmx_palettes(),
  scale.percent = TRUE
)
}
\arguments{
\item{odata}{the observed data with one row per subject with a time and event indicator (1==event, 0==censored)}

\item{sdata}{the simulated data with one row per subject and replicate with a time and event indicator (1==event, 0==censored)}

\item{time}{what indicates event/censoring time}

\item{event}{what indicates if an event occurred}

\item{iter}{what indicates the replicate number}

\item{strat}{character vector: what stratifications need to be applied}

\item{ci}{what level of confidence to we want for the simulated Kaplan-Meier}

\item{cuminc}{do we want to show the inverse Kaplan-Meier (1-KM) then set to TRUE, for survival-like set to FALSE}

\item{simCol}{color for the simulation interval}

\item{obsCol}{color for the observed line}

\item{posObs}{provide an explicit set of times to smoothen the CI for simulations}

\item{show.censor}{logical to indicate whether the censoring times should be shown for the observed}

\item{censor.size}{}
}
\description{
Title
}
