% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kaplan_Meier_curves.R
\name{Kaplan_Meier_curves}
\alias{Kaplan_Meier_curves}
\title{Title}
\usage{
Kaplan_Meier_curves(data_frame_name)
}
\arguments{
\item{RTTEdata}{dataframe for analysis}
}
\value{
Kaplan-Meier (KM) curves for the provided data coloured by treatment/dose
}
\description{
Title
}
