% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_kaplanmeier_utils.R
\name{kpm}
\alias{kpm}
\alias{kpm_se}
\alias{kpm_ci}
\title{Kaplan-Meier estimator (low-level implementation)}
\usage{
kpm(x)

kpm_se(x)

kpm_ci(x, width = 0.95)
}
\arguments{
\item{x}{a vector of numeric, with possible values of \code{0} (censoring) or \code{1} (event), or coercible to this definition.}

\item{width}{a numeric, width of the confidence interval.}
}
\value{
\item{\code{kpm()}}{a vector of numeric, Kaplan-Meier estimator of the survival}
\item{\code{kpm_se()}}{a vector of numeric, standard error associated to the
Kaplan-Meier estimator of the survival, as returned by \code{survival::survfit()}}
\item{\code{kpm_ci()}}{a two-column numerical matrix, confidence interval
associated to the Kaplan-Meier estimator of the survival, as returned by
\code{survival::survfit()} with default settings.}
}
\description{
Kaplan-Meier estimator (low-level implementation)
}
\examples{
x <- c(1,1,1,0,1,0)
kpm(x)
kpm_se(x)
kpm_ci(x, width = .95)

# Errors if x does not have a valid format
\dontrun{
kpm(c(1,2,1,2,1))
kpm(c("Event", "No event"))
}
}
