% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_kaplanmeier_utils.R
\name{kpm_stepfun}
\alias{kpm_stepfun}
\alias{kpm_ci_stepfun}
\alias{atrisk_stepfun}
\title{Kaplan-Meier step functions}
\usage{
kpm_stepfun(time, dv, cuminc = FALSE)

kpm_ci_stepfun(time, dv, cuminc = FALSE, width = 0.95)

atrisk_stepfun(time)
}
\arguments{
\item{time}{a vector of non-missing numeric values, time of the observation}

\item{dv}{a vector of non-missing numeric values, type of observation
(\code{0} (censoring) or \code{1} (event))}

\item{cuminc}{a logical. If \code{TRUE}, calculates the cumulative incidence of
events. If \code{FALSE} (the default), the survival function.}

\item{width}{a numeric, width of the confidence interval.}
}
\value{
step functions (see \link[stats:stepfun]{stats::stepfun}):
\item{\code{kpm_stepfun()}}{a step function that calculates the survival
(or cumulative incidence) as a function of time according to the
Kaplan-Meier estimator of survival.}
\item{\code{kpm_ci_stepfun()}}{a step function that calculates the confidence
interval of the survival (or cumulative incidence) as a function of time
according to the Kaplan-Meier estimator of survival and the Greenwood's
formula.}
\item{\code{kpm_atrisk()}}{a step function that calculates the number of
subjects at risk as a function of time.}
}
\description{
The Kaplan-Meier estimator only provides an estimation of the survival at the
times available in the data. Often, we want to interpolate the survival at
any time points, with the following rules:
\itemize{
\item Survival at time 0 is equal to 1.
\item Survival between two time points is equal to the survival at the
previous time point (the survival value is carried forward).
}

The functions documented here create step functions that reproduces these
interpolation rules, and can predict survival at any provided time.
}
\examples{
# Times do not have to be sorted
# (as in a regular data set where data are stored per subject identifier)

data <- data.frame(
  TIME = c(24, 5, 58, 24, 12, 36, 24),
  DV   = c( 1, 1,  0,  1,  0,  1,  1)
)

KPM <- kpm_stepfun(time = data$TIME, dv = data$DV)
KPMCI <- kpm_ci_stepfun(time = data$TIME, dv = data$DV)
ATRISK <- atrisk_stepfun(time = data$TIME) # no argument `dv`
# KPM, KPMCI and ATRISK are functions that only takes one argument

outtimes <- seq(0, 60, 12) # Times we want the survival calculated

KPM(outtimes)
KPMCI(outtimes)
ATRISK(outtimes)

}
