% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummaryTTE.R
\name{makeSummaryTableTTE}
\alias{makeSummaryTableTTE}
\title{Title}
\usage{
makeSummaryTableTTE(
  df,
  myID = "ID",
  myDV = "DV",
  digits = 3,
  nObsColNm = "\\\\textbf{nEvent\\\\textsuperscript{b}}",
  avnObsColNm = "\\\\textbf{pEvent/nID\\\\textsuperscript{c}}",
  caption = "Number of patients and number of events",
  footnote =
    "\\\\textsuperscript{a}Number of patients\\\\newline\\\\textsuperscript{b}Number of events\\\\newline\\\\textsuperscript{c}Proportion of number of events",
  myFun = NULL,
  ...
)
}
\arguments{
\item{nEventColNm}{the charactrer string to be printed as the column name with number of events}

\item{avnEventColNm}{the charactrer string to be printed as the column name with average number of events}
}
\description{
Title
}
\examples{
ttedata  <- readr::read_csv(system.file('extdata/anaDATAf.csv', package= 'PMXtte'), show_col_types = FALSE)

makeSummaryTableTTE(ttedata,
                    outerLevel   = "DOSEN",
                    outerLabel   = "Dose",
                    innerLevel   = "STUDYIDN",
                    innerLabel   = "Study",
                    asList = TRUE)

#output as latex table
makeSummaryTableTTE(ttedata,
                    outerLevel   = "DOSEN",
                    outerLabel   = "Dose",
                    innerLevel   = "STUDYIDN",
                    innerLabel   = "Study")
}
