% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhRame.R
\name{PhRame_makeSummaryTable}
\alias{PhRame_makeSummaryTable}
\title{Single data summary table including all strata}
\usage{
PhRame_makeSummaryTable(
  df,
  addFactors = FALSE,
  filterExpr = NULL,
  myID = "ID",
  myDV = "DV",
  outerLevel = NULL,
  innerLevel = NULL,
  outerLabel = NULL,
  innerLabel = NULL,
  digits = 3,
  nIdColNm = "\\\\textbf{nID\\\\textsuperscript{a}}",
  nObsColNm = "\\\\textbf{nObs\\\\textsuperscript{b}}",
  avnObsColNm = "\\\\textbf{nObs/nID\\\\textsuperscript{c}}",
  caption =
    "PK analysis data set: Number of subjects with observations and number of observations",
  label = "tab:anaSummary",
  footnote =
    "\\\\textsuperscript{a}Number of subjects\\\\newline\\\\textsuperscript{b}Number of observations\\\\newline\\\\textsuperscript{c}Average number of observations per subject",
  footnoteSize = "footnotesize",
  textSize = "small",
  here = TRUE,
  numeric.dollar = FALSE,
  asList = FALSE,
  saveTex = FALSE,
  fileName = NULL,
  filePath = "./",
  myFun = NULL,
  ...
)
}
\arguments{
\item{df}{is the data frame (e.g. anaDataF) or data object created using
\code{pmxdata::derived_data} (e.g. DATobject)}

\item{addFactors}{if \strong{TRUE}, factors are added to the input data.
Default is \strong{FALSE}. This option is used iff \code{df} is a data
object}

\item{filterExpr}{is an expression used to filter the data. Either a formula
using non standard evaluation (e.g. \code{quo(AMT==0 & TYPE==1)}) or a
character string with filter expression (e.g. "(AMT==0 & TYPE==1)") is
acceptable to this argument.}

\item{myID}{is the name of the ID column, default is \strong{"ID"}}

\item{myDV}{is the name of the DV column, default is \strong{"DV"}}

\item{outerLevel}{is the outler level (1st level) of stratification variable,
eg. "STUDYF"}

\item{innerLevel}{is the inner level (2nd level) of stratification variable,
eg. "DOSEF"}

\item{outerLabel}{is the label for the outler level (1st level) of
stratification variable, eg. "Study")}

\item{innerLabel}{is the label for the inner level (2nd level) of
stratification variable, eg. "Dose"}

\item{digits}{is the number of significant digits}

\item{nIdColNm}{is the character string to be printed as the column name
with the number of subjects, default is
\strong{"\\\\textbf\{nID\\\\textsuperscript\{a\}\}"}}

\item{nObsColNm}{is the character string to be printed as the column name
with the number of observations, default is
\strong{"\\\\textbf\{nObs\\\\textsuperscript\{b\}\}"}}

\item{avnObsColNm}{is the character string to be printed as the column name
with the average number of observations per subject in a given strata,
default is
\strong{"\\\\textbf\{nObs/nID\\\\textsuperscript\{c\}\}"}}

\item{caption}{is the table caption. Assign \strong{NULL} to this argument
produce table without caption. Default is \strong{"PK analysis data set:
Number of subjects with observations and number of observations"}}

\item{label}{is the label for the table used for cross-reference, default is
\strong{"tab:anaSummary"}}

\item{footnote}{is the text for footnote, default is
\strong{"\\textsuperscript\{a\} Number of
subjects\\\\newline\\\\textsuperscript\{b\} Number of
observations\\\\newline\\\\textsuperscript\{c\} Average number of
observations per subject"}}

\item{footnoteSize}{is the footnote text size, default is
\strong{"footnotesize"}}

\item{textSize}{is the text size, default is \strong{"small"}}

\item{here}{is the logical argument determining the placement of the LaTex
table. If set to \strong{TRUE} with optional argument \code{table.env=TRUE}
with \code{longtable=FALSE}, will cause the LaTeX table environment to be
set up with option H to guarantee that the table will appear exactly where
you think it will in the text. Default is \strong{TRUE}}

\item{numeric.dollar}{is the logical argument determining if the numbers will
be placed within $ symbols, default is \strong{FALSE}}

\item{asList}{if \strong{TRUE} the data summary is returned as a list where
different elements in the list contain summary of different stratification
levels. If \strong{FALSE} the data summary is returned as a latex table. If
no stratification variable is provided, this function always returns a list
of two items (Number of subjects, Number of observarions and Average number
of observarions per subject)}

\item{saveTex}{is the logical argument determining if the LaTex code for the
table will be saved on disk. Default is \strong{FALSE}}

\item{fileName}{is the name of the Tex file (without extension) with the
LaTex code for the table saved on disk, eg. if \code{fileName="mytable"} is
provided, a file called \code{"mytable.tex"} will be saved on disk. Default
is \strong{NULL}.}

\item{filePath}{is the path to the directory where the Tex file is saved.
Default is \strong{"/"}.}

\item{myFun}{internal function for the calculation of summarized data. If
NULL, the default, an internal function is used. This should not be changed
for a standard use.}

\item{...}{is the additional optional arguments compatible with
\link[Hmisc:latex]{Hmisc::latex}}
}
\description{
Function to derive a single data summary table for all strata
(upto 2 levels of stratification) that can be used to format for latex
table
}
