% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKaplanMeier.R
\name{plotKaplanMeier}
\alias{plotKaplanMeier}
\title{Create Kaplan Meier plots}
\usage{
plotKaplanMeier(
  data,
  time_col = "TSFDW",
  event_col = "DV",
  cov_col = NULL,
  facet.by = NULL,
  panel.labs = NULL,
  short.panel.labs = TRUE,
  label.parsed = TRUE,
  xlab = "Time since first dose (weeks)",
  ylab = "Subjects without events (\%)",
  break.time.by = 4,
  xlim = NULL,
  ylim = c(0, 1),
  surv.scale = "percent",
  ggtheme = PMXColors_theme_pmx(),
  palette = PMXColors_pmx_palettes(),
  conf.int = TRUE,
  ciWidth = 0.95,
  add.ciWidth.to.legend = TRUE,
  conf.int.alpha = 0.3,
  conf.int.fill = "gray",
  risk.table = TRUE,
  risk.table.y.text = TRUE,
  risk.table.y.text.col = TRUE,
  risk.table.fontsize = 8,
  table.clip = "off",
  pval = FALSE,
  pval.method = TRUE,
  pval.size = 10,
  surv.median.line = "hv",
  surv.median.line.legend = TRUE,
  surv.median.line.title = "Median Survival",
  legend = "top",
  legend.title = "Observed",
  legend.nrow = NULL,
  legend.ncol = NULL,
  risk.table.ylab = cov_col,
  risk.table.title = "Number of subjects at risk",
  legend.labs = sub(pattern = paste(cov_col, "=", sep = ""), replacement = "", x =
    names(fit$strata)),
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  legend.box = "horizontal",
  legend.title.size = 9,
  legend.spacing = 5,
  censor.shape = "|",
  censor.size = 4,
  ...
)
}
\arguments{
\item{data}{dataframe used to fit survival curves}

\item{time_col}{The name of the column containing the survival times.}

\item{event_col}{The name of the column indicating an event occurrence (1) or censoring (0)}

\item{cov_col}{The name of the column to stratify on, must be string (e.g. "DOSE") if NULL there will be no stratification}

\item{facet.by}{a character vector containing the name of grouping variables to facet the survival curves into multiple panels. Should be of length <= 2. Alias of the ggsurvplot_facet() function. the function does not facet tables therefor only the plots will be faceted and tables will not be produced.}

\item{panel.labs}{a list of one or two character vectors to modify facet label
text. For example, panel.labs = list(sex = c("Male", "Female")) specifies
the labels for the "sex" variable. For two grouping variables, you can use
for example panel.labs = list(sex = c("Male", "Female"), rx = c("Obs",
"Lev", "Lev2") ).}

\item{short.panel.labs}{logical value. Default is FALSE. If TRUE, create short
labels for panels by omitting variable names; in other words panels will be
labelled only by variable grouping levels.}

\item{label.parsed}{if TRUE (default), facet labels can include superscipts and greek letters}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{break.time.by}{numeric value controlling time axis breaks. Default value
is NULL.}

\item{xlim}{x axis limits e.g. \code{xlim = c(0, 1000)}.}

\item{ggtheme}{function, ggplot2 theme name. Default value is theme_pmx. Allowed values include ggplot2 official themes.}

\item{palette}{the color palette to be used. Allowed values include "hue" for
the default hue color scale; "grey" for grey color palettes; brewer palettes
e.g. "RdBu", "Blues", ...; or custom color palette e.g. c("blue", "red"); and scientific journal palettes from ggsci R package, e.g.: "npg",
 "aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and
 "rickandmorty".
See details section for more information.  Can be also a numeric vector of
length(groups); in this case a basic color palette is created using the
function \link[grDevices]{palette}.}

\item{conf.int}{logical value. If TRUE, plots confidence interval.}

\item{ciWidth}{confidence intervals to be passed in surv_fit default is 0.95}

\item{risk.table}{Allowed values include: \itemize{ \item TRUE or FALSE
specifying whether to show or not the risk table. Default is TRUE \item
"absolute" or "percentage". Shows the \bold{absolute number} and the
\bold{percentage} of subjects at risk by time, respectively. \item "abs_pct"
to show both absolute number and percentage. \item "nrisk_cumcensor" and
"nrisk_cumevents". Show the number at risk and, the cumulative number of
censoring and events, respectively. }}

\item{risk.table.fontsize}{size of the font for survival table in points}

\item{pval}{logical value, a numeric or a string. If logical and TRUE, the p-value is added on the plot. If numeric, then the computed p-value is substituted with the one passed with this parameter. If character, then the customized string appears on the plot.}

\item{pval.method}{whether to add a text with the test name used for
calculating the pvalue, that corresponds to survival curves' comparison -
used only when \code{pval=TRUE}}

\item{pval.size}{size of font for the pval in points}

\item{surv.median.line}{character vector for drawing a horizontal/vertical
line at median survival. Allowed values include one of c("none", "hv", "h",
"v"). v: vertical, h:horizontal.}

\item{surv.median.line.legend}{Text to be used for median survival line in the legend}

\item{legend}{character specifying legend position. Allowed values are one of
c("top", "bottom", "left", "right", "none"). Default is "top" side position.
to remove the legend use legend = "none". Legend position can be also
specified using a numeric vector c(x, y); see details section.}

\item{legend.title}{legend title.}

\item{legend.nrow}{The number if rows in the legend for the stratification variable}

\item{legend.ncol}{The number if columns in the legend for the stratification variable}

\item{risk.table.title}{The title to be used for the risk table.}

\item{legend.labs}{character vector specifying legend labels. Used to replace
the names of the strata from the fit. Should be given in the same order as
those strata.}

\item{nrow, ncol}{Number of rows and columns in the pannel. Used only when the
data is faceted by one grouping variable.}

\item{scales}{should axis scales of panels be fixed ("fixed", the default),
free ("free"), or free in one dimension ("free_x", "free_y").}

\item{...}{
  Arguments passed on to \code{\link[survminer:ggsurvplot]{survminer::ggsurvplot}}
  \describe{
    \item{\code{fit}}{allowed values include:
\itemize{
\item a survfit object
\item a list of survfit objects. Passed to \code{\link[survminer]{ggsurvplot_list}()}
\item a data frame containing survival curves summary. Passed to \code{\link[survminer]{ggsurvplot_df}()}.
}}
    \item{\code{fun}}{an arbitrary function defining a transformation of the survival
curve.  Often used transformations can be specified with a character
argument: "event" plots cumulative events (f(y) = 1-y), "cumhaz" plots the
cumulative hazard function (f(y) = -log(y)), and "pct" for survival
probability in percentage.}
    \item{\code{color}}{color to be used for the survival curves. \itemize{ \item If the
number of strata/group (n.strata) = 1, the expected value is the color name.
For example color = "blue". \item If n.strata > 1, the expected value is the
grouping variable name. By default, survival curves are colored by strata
using the argument color = "strata",  but you can also color survival curves
by any other grouping variables used to fit the survival curves. In this
case, it's possible to specify a custom color palette by using the argument
palette.}}
    \item{\code{linetype}}{line types. Allowed values includes i) "strata" for changing
linetypes by strata (i.e. groups); ii) a numeric vector (e.g., c(1, 2)) or a
character vector c("solid", "dashed").}
    \item{\code{test.for.trend}}{logical value. Default is FALSE. If TRUE, returns the
test for trend p-values. Tests for trend are designed to detect ordered
differences in survival curves. That is, for at least one group. The test
for trend can be only performed when the number of groups is > 2.}
    \item{\code{cumevents}}{logical value specifying whether to show or not the table of
the cumulative number of events. Default is FALSE.}
    \item{\code{cumcensor}}{logical value specifying whether to show or not the table of
the cumulative number of censoring. Default is FALSE.}
    \item{\code{tables.height}}{numeric value (in [0 - 1]) specifying the general height
of all tables under the main survival plot.}
    \item{\code{group.by}}{a character vector containing the name of grouping variables. Should be of length <= 2.
Alias of the \code{\link[survminer]{ggsurvplot_group_by}()} function.}
    \item{\code{add.all}}{a logical value. If TRUE, add the survival curve of pooled patients (null model) onto the main plot.
Alias of the \code{\link[survminer]{ggsurvplot_add_all}()} function.}
    \item{\code{combine}}{a logical value. If TRUE, combine a list survfit objects on the same plot.
Alias of the \code{\link[survminer]{ggsurvplot_combine}()} function.}
    \item{\code{tables.theme}}{function, ggplot2 theme name. Default value is
\link[survminer]{theme_survminer}. Allowed values include ggplot2 official themes: see
\code{\link[ggplot2]{theme}}. Note that, \code{tables.theme} is incremental to \code{ggtheme}.}
  }}
}
\value{
A "ggsurvplot" object, see the documentation of \code{survminer::ggsurvplot} for more information. Basically it is a list that notably contains two "ggplot" objects, "x$plot" and "x$table", that renders into a single figure the Kaplan-Meier curves on top and the risk table on the bottom.
}
\description{
This function is deprecated. Please use \code{ggKAP()} instead.
}
\details{
This function generates Kaplan Meier curves plot using ggsurvplot from a dataframe

The function takes a dataframe (tibble will be coerced to data.frame) and fits the survival curves using surv_fit then plots them using ggsurvplot.
The arguments risk.table.fontsize and pval.size convert regular sizes to ggplot sizes. For other font size arguments passed to ggsurvplot you can multiply regular point sizes by 0.36 to convert to ggplot sizes. To change the color of the curves use the argument palette instead of color.
}
\examples{
library(magrittr)
#load example data
rttedata <- readr::read_csv(system.file('extdata/DAT-1c-RED-1a-PMX-WOWTTE-PFPMX-1.csv', package= 'PMXtte'), show_col_types = FALSE)
rttedata <- dplyr::filter(rttedata, EVID == 0, TYPE == 2)

#generate time to first event data
tte1data <- rttedata \%>\% filter_xth_event(1)

#transform back to "data.frame" not "tibble" do avoid bugs
tte1data <- as.data.frame(tte1data)

#default plot with one curve
plotKaplanMeier(tte1data)

# Kaplan Meier plot stratified by Dose and faceted by SEX
plotKaplanMeier(tte1data, cov_col = "DOSEN", facet.by = 'SEXN')

# Change the panel labs of the facets
plotKaplanMeier(tte1data, cov_col = 'DOSEN', facet.by = 'SEXN', panel.labs = list('SEXN' = c('male','female')))

# Change the size of pval text
plotKaplanMeier(tte1data, cov_col = "DOSEN", pval = TRUE, pval.size = 12)

# Change the starting color in the palette
plotKaplanMeier(tte1data, cov_col = "DOSEN", palette = PMXtte:::PMXColors_pmx_palettes(firstColorNum = 3))

# Change the color for a single curve plot
plotKaplanMeier(tte1data, palette = 'cadetblue')

# Create custom palette of colors
Colors <- c('cadetblue', 'coral1', 'darkorchid', 'deepskyblue3', 'darkred')
plotKaplanMeier(tte1data, palette = Colors, cov_col = 'DOSEN')

#modify plot using ggplot syntax
p <- plotKaplanMeier(tte1data, cov_col = "DOSEN")
p$table <- p$table + ggplot2::ylab('Dose (mg)')
p

# remove median line from legend
plotKaplanMeier(tte1data, cov_col = "DOSEN", surv.median.line.legend = FALSE)

# Time to second event data
# !! NOT ADVISED since the subjects at risk are not at random across groups!!
rttedata \%>\%
  filter_xth_event(2) \%>\%
  plotKaplanMeier( # KPM plot of time to second event data
    cov_col = "DOSEN",
    time_col = "TSLE", #x-axis: time since Last event (not first dose),
    xlab = "Time since last event (week)" #in week because `time_col="TSFDW"`
  )

}
