% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dataframe.R
\name{prep_dataframe}
\alias{prep_dataframe}
\title{Title}
\usage{
prep_dataframe(RTTEdata, plottype = "time_to_first_event")
}
\arguments{
\item{RTTEdata}{dataframe for to be prepared for plots}

\item{plottype}{string selecting type of plot}
}
\value{
modified dataframe ready for plots
}
\description{
Title
}
