% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simplettedata}
\alias{simplettedata}
\title{Dummy "observed" Time-To-Event data}
\format{
\subsection{\code{simplettedata}}{

A data frame with 50 rows and 5 columns:
\describe{
\item{ID}{a numeric, the subject identifier (from 0 to 50)}
\item{TIME}{a numeric, the time of the observation (from 0 to 100, rounded to closest integer)}
\item{DV}{a numeric, the dependent variable, 0 for censoring, 1 for event}
\item{SEXF}{a factor, with two levels: "Male" and "Female"}
\item{AUCQF}{a factor, with three levels: "1st", "2nd" and "3rd"}
...
}
}
}
\usage{
simplettedata
}
\description{
Dummy "observed" Time-To-Event data
}
\keyword{datasets}
