% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_kaplanmeier_ggproto.R
\name{stat_kaplanmeier_pval}
\alias{stat_kaplanmeier_pval}
\title{P-value of the likelihood ratio test between several Kaplan-Meier functions}
\usage{
stat_kaplanmeier_pval(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...,
  xpos = NULL,
  ypos = NULL
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{xpos, ypos}{numeric values, the coordinates of the text. By default (NULL),
coordinates that corresponds toon the bottom left corner will be computed.}
}
\description{
P-value of the likelihood ratio test between several Kaplan-Meier functions
}
\examples{
library(ggplot2)
dat <- PMXtte::simplettedata

# Requires a `strata` aesthetic argument
ggplot(dat) +
  stat_kaplanmeier_pval(aes(x = TIME, y = DV, strata = SEXF))

# Better if supported with the survival function
# Make sure the grouping is not done in the main `ggplot()` function.
# If you do so, p-value will attempted to be calculated _within_ each group
# However we want the p-value to be calculated _between_ each strata level
ggplot(dat, aes(x = TIME, y = DV)) +
  stat_kaplanmeier(aes(group = AUCQF)) +
  stat_kaplanmeier_pval(aes(strata = AUCQF))

# Supported facetting
# P-value calculated for each panel
ggplot(dat, aes(x = TIME, y = DV)) +
  stat_kaplanmeier(aes(group = AUCQF)) +
  stat_kaplanmeier_pval(aes(strata = AUCQF)) +
  facet_grid(~SEXF)

# May not support coloring or any variable-related aesthetic mapping
# Despite that, the guide (legend) will carry some text attributes
# P-value calculated for each panel
ggplot(dat, aes(x = TIME, y = DV)) +
  stat_kaplanmeier(aes(group = AUCQF, col = SEXF)) +
  stat_kaplanmeier_pval(aes(strata = AUCQF, col = SEXF)) + # not colored
  facet_grid(~SEXF)

# No argument needed if the cumulative incidence is plotted.
ggplot(dat, aes(x = TIME, y = DV)) +
  stat_kaplanmeier(aes(group = AUCQF), cuminc = TRUE) +
  stat_kaplanmeier_pval(aes(strata = AUCQF))

# The position of the text is defaulted to the bottom-left of the panel
# Change it with `xpos` and `ypos`
ggplot(dat, aes(x = TIME, y = DV)) +
  stat_kaplanmeier(aes(group = AUCQF), cuminc = TRUE) +
  stat_kaplanmeier_pval(aes(strata = AUCQF), xpos = 25, ypos = .75)

# Default geom is "text", any compatible option can be used
ggplot(dat, aes(x = TIME, y = DV)) +
  stat_kaplanmeier(aes(group = AUCQF)) +
  stat_kaplanmeier_pval(aes(strata = AUCQF), colour = "red", size = 10)

# or change the geom
ggplot(dat, aes(x = TIME, y = DV)) +
  stat_kaplanmeier(aes(group = AUCQF)) +
  stat_kaplanmeier_pval(
    aes(strata = AUCQF),
    geom = "label", color = "red", fill = "yellow"
  )
}
