% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_kaplanmeier_ggproto.R
\docType{data}
\name{StatKaplanMeierGGproto}
\alias{StatKaplanMeierGGproto}
\alias{StatKaplanMeier}
\alias{StatKaplanMeierCensor}
\alias{StatKaplanMeierMedian}
\alias{StatKaplanMeierSE}
\alias{StatKaplanMeierPval}
\alias{StatKaplanMeierSim}
\alias{StatKaplanMeierRiskTable}
\title{StatKaplanMeier ggproto objects}
\format{
An object of class \code{StatKaplanMeier} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{StatKaplanMeierCensor} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatKaplanMeierMedian} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatKaplanMeierSE} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{StatKaplanMeierPval} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{StatKaplanMeierSim} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{StatKaplanMeierRiskTable} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatKaplanMeier

StatKaplanMeierCensor

StatKaplanMeierMedian

StatKaplanMeierSE

StatKaplanMeierPval

StatKaplanMeierSim

StatKaplanMeierRiskTable
}
\description{
\code{StatKaplanMeier}
\code{StatKaplanMeierCensor}
\code{StatKaplanMeierMedian}
\code{StatKaplanMeierSE}
\code{StatKaplanMeierPval}
\code{StatKaplanMeierSim}
\code{StatKaplanMeierRiskTable}
}
\seealso{
\code{\link[ggplot2:geom_path]{ggplot2::geom_step()}}, \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}},
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}, \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}}, \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} from
the ggplot2 package.
}
\keyword{datasets}
