% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_kaplanmeier_ggproto.R
\docType{data}
\name{StatKaplanMeierSE}
\alias{StatKaplanMeierSE}
\title{Standard error-based confidence interval on the survival function}
\format{
An object of class \code{StatKaplanMeierSE} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatKaplanMeierSE
}
\arguments{
\item{cuminc}{a logical. If \code{TRUE}, calculates the cumulative incidence of events. If \code{FALSE} (the default), the survival function.}

\item{level}{Level of confidence interval to use (0.95 by default)}
}
\description{
Standard error-based confidence interval on the survival function
}
\examples{
library(ggplot2)
dat <- PMXtte::simplettedata

# Is of little value on itself...
ggplot(dat) +
  stat_kaplanmeier_se(aes(x = TIME, y = DV))

# Better if supported with the survival function
ggplot(dat, aes(x = TIME, y = DV)) +
  stat_kaplanmeier_se(fill = "grey60") +
  stat_kaplanmeier()

# Supported variables defined in `aes()` and facetting
ggplot(dat, aes(x = TIME, y = DV, col = AUCQF, fill = AUCQF)) +
  stat_kaplanmeier_se(alpha = .4) +
  stat_kaplanmeier() +
  facet_grid(~SEXF)

# Display the cumulative incidence
ggplot(dat, aes(x = TIME, y = DV)) +
  stat_kaplanmeier_se(cuminc = TRUE, fill = "grey60") +
  stat_kaplanmeier(cuminc = TRUE)

# Default geom is "ribbon", any compatible option can be used
ggplot(dat, aes(x = TIME, y = DV)) +
  stat_kaplanmeier_se(linetype = 3, size = 5, color = "red", fill = "yellow")+
  stat_kaplanmeier()

# or change the geom
ggplot(dat, aes(x = TIME, y = DV)) +
  stat_kaplanmeier_se(geom = "errorbar", color = "red") +
  stat_kaplanmeier()
}
\keyword{datasets}
