% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryCountRTTE.R
\name{summaryCountRTTE}
\alias{summaryCountRTTE}
\title{Single data summary table for (R)TTE data}
\usage{
summaryCountRTTE(
  df,
  addFactors = FALSE,
  filterExpr = NULL,
  myID = "ID",
  myDV = "DV",
  myEVCOUNT = "EVCOUNT",
  outerLevel = NULL,
  innerLevel = NULL,
  outerLabel = NULL,
  innerLabel = NULL,
  digits = 3,
  lumpCount = Inf,
  dropCount0 = FALSE,
  nIdColNm = "\\\\textbf{nID\\\\textsuperscript{a}}",
  nObsColNm = NULL,
  caption =
    "PK analysis data set: Number of subjects with observations and number of events",
  label = "tab:anaSummary",
  footnote = "\\\\textsuperscript{a}Number of subjects",
  footnoteSize = "footnotesize",
  textSize = "small",
  here = TRUE,
  numeric.dollar = FALSE,
  asList = FALSE,
  saveTex = FALSE,
  fileName = NULL,
  filePath = "./",
  myFun = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Single data summary table for (R)TTE data
}
