% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTTESim.R
\name{update_sub_advan_tol}
\alias{update_sub_advan_tol}
\title{Update NONMEM $SUBROUTINES to ADVAN=6 TOL=6}
\usage{
update_sub_advan_tol(linesSub)
}
\arguments{
\item{linesSub}{A string containing the $SUBROUTINES line}
}
\value{
A string with ADVAN=6 and TOL=6 enforced
}
\description{
Update NONMEM $SUBROUTINES to ADVAN=6 TOL=6
}
