% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_xth_event.R
\name{filter_xth_event}
\alias{filter_xth_event}
\title{Filter time to Xth event data}
\usage{
filter_xth_event(
  x,
  event = 1,
  event_col = "DV",
  evcount_col = "EVCOUNT",
  time_since_last_event = "TSLE",
  time_col = "TSFDW"
)
}
\arguments{
\item{x}{a dataset, with only observations regarding time to event data. It often means data filtered with \code{EVID==0} and with a unique "TYPE" value.}

\item{event}{a integer of length 1, informs if the time to 1st, 2nd, 3rd event (etc...) should be returned. Default is 1 (time to first event data).}

\item{event_col}{a character, the event column in the data.}

\item{evcount_col}{a character, the event count column in the data. Expected as the cumulative sum of the event column (i.e. starts at 0, not incremented of +1 in case of censoring event).}

\item{time_since_last_event}{a character, the name of the column to create that will inform the time since the last event. If \code{NULL}, no column will be created. Ignored if \code{event=1}.}

\item{time_col}{a character, the name of the time column used to calculate the time since last event. Ignored if \code{time_since_last_event=NULL} or if \code{event=1}.}
}
\value{
a dataset, filtered for the analysis of a particular event number (time to 1st event, time to 2nd event, etc...). If the event number (\code{event}) is greater than 1, a new column that informs the time since the last event is created (default is \code{TSLE}, based on \code{TSFDW}).
}
\description{
Filter a repeated time to event dataset into a time to 1st event, time to 2nd event (etc...) dataset. Particularly useful for summary statistics, Kaplan-Meier plots or Visual Predictive checks of time to second (and more) event data.
}
\examples{
library(magrittr)
# Read data
rttedata <- readr::read_csv(system.file('extdata/DAT-1c-RED-1a-PMX-WOWTTE-PFPMX-1.csv', package= 'PMXtte'), show_col_types = FALSE)
# Filter observations
rttedata <- dplyr::filter(rttedata, EVID == 0, TYPE == 2)
filter_xth_event(rttedata) # By default, filters event 1
filter_xth_event(rttedata, event = 2)
filter_xth_event(rttedata, event = 3) # etc... # The number of ID decreases

rttedata \%>\%
  filter_xth_event(2) \%>\%
  plotKaplanMeier( # KPM plot of time to second event data
    cov_col = "DOSEN",
    time_col = "TSLE", #x-axis: time since Last event (not first dose),
    xlab = "Time since last event (week)" #in week because `time_col="TSFDW"`
  )

}
