% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggKAP.R
\name{ggKAP}
\alias{ggKAP}
\title{Visualize Survival Data}
\usage{
ggKAP(
  data,
  sdata,
  time_var = "TIME",
  dv_var = "DV",
  iter_var = "ITER",
  color_var = NULL,
  facet_var = NULL,
  show_kpm = !missing(data),
  show_censor = !missing(data),
  show_risktable = !missing(data),
  show_se = !missing(data) && missing(sdata),
  show_pval = !missing(data) && missing(sdata) && !is.null(color_var) && (if
    (!is.null(facet_var)) {
     color_var != facet_var
 } else {
     TRUE
 }),
  show_median = !missing(data) && missing(sdata),
  show_sim = !missing(sdata),
  cuminc = FALSE,
  ci_level = if (show_se) 0.95 else 0.9,
  ci_alpha = 0.5,
  censor_shape = "|",
  censor_size = 4,
  median_linetype = 2,
  pval_pos = NULL,
  scale_x_breaks = ggplot2::waiver(),
  scale_x_break_by = NULL,
  scale_y_labels = scales::percent,
  scale_y_risktable_reverse = FALSE,
  label_x = "Time",
  label_y_fig = paste0("Subjects with", if (!cuminc) {
     "out"
 }, " event (\%)"),
  label_color = color_var,
  label_y_risktab = label_color,
  legend_label_kpm = "Observed",
  legend_label_censor = "Censored",
  legend_label_se = paste0(scales::percent(ci_level), " CI"),
  legend_label_median = "Median Survival",
  legend_label_sim = paste0("Simulated ", scales::percent(ci_level), " CI"),
  guide_legend_nrow = NULL,
  scale_color_labels = waiver(),
  scale_color_values = unname(PMXColors_PMXcolors$default),
  facetting_args = list(),
  xlim = c(NA_real_, NA_real_),
  ylim = c(0, 1),
  title_risktable = "Number of subjects at risk",
  arrange = TRUE,
  rel_heights = list(legends = c(1, 1), figtable = c(2, 1), overall = c(1, 8)),
  check_input = TRUE
)
}
\arguments{
\item{data}{a data.frame, observed data, assuming one time-to-event record per individual.}

\item{sdata}{a data.frame, simulated data, assuming one time-to-event record per individual and iteration.}

\item{time_var}{a character, the name of the time column in \code{data} and \code{sdata}. The variable should be numeric, ideally > 0.}

\item{dv_var}{a character, the name of the dependent variable column in \code{data} and \code{sdata}. The variable should be numeric, either 0 or 1.}

\item{iter_var}{a character, the name of the iteration column in \code{sdata}. The variable should be numeric, ideally a sequence of integers.}

\item{color_var}{a character, the name of the column in \code{data} and \code{sdata} used to stratify by color. The variable should preferably be a factor, possibly a character or a logical, but not a double or integer.}

\item{facet_var}{a character, the name of the column in \code{data} and \code{sdata} used to facet the plot (Kaplan-Meier figure and risk tables) . The variable should preferably be a factor, possibly a character or a logical, but not a double or integer.}

\item{show_kpm}{a logical, should the Kaplan-Meier curve (and associated legend) appear on the plot? Default is \code{TRUE} if \code{data} is provided.}

\item{show_censor}{a logical, should the censored events (and associated legend) appear on the plot? Default is \code{TRUE} if \code{data} is provided.}

\item{show_risktable}{a logical, should the table with number of patient at risk appear on the plot? Default is \code{TRUE} if \code{data} is provided.}

\item{show_se}{a logical, should a parametric confidence interval (and associated legend) around the Kaplan-Meier curve (computed from observed data) appear on the plot? Default is \code{TRUE} if \code{data} alone is provided, \code{FALSE} if \code{sdata} is also provided.}

\item{show_pval}{a logical, should the p-value of a log-rank test appear on the plot? Default is \code{TRUE} if \code{data} alone is provided, and there are several Kaplan-Meier curves in each panel of the plot.}

\item{show_median}{a logical, should the median survival (and associated legend) appear on the plot? Default is \code{TRUE} if \code{data} alone is provided, \code{FALSE} if \code{sdata} is also provided.}

\item{show_sim}{a logical, should a confidence interval (and associated legend) computed from simulated data (\code{sdata}) appear on the plot? Default is \code{TRUE} if \code{sdata} is provided.}

\item{cuminc}{a logical, should the plot represent the cumulative incidence? If \code{FALSE}, the default, the survival is represented.}

\item{ci_level}{a numeric, between 0 and 1, indicating the level of the confidence interval. Default is \code{.95} (95\%) if the confidence interval is calculated from observed data (\code{data}), \code{.90} (90\%) if calculated from simulations (\code{sdata}).}

\item{ci_alpha}{a numeric, between 0 and 1, indicating the opacity of the confidence interval.}

\item{censor_shape}{shape of the censoring event.}

\item{censor_size}{a numeric, the size of the censoring event.}

\item{median_linetype}{a numeric, linetype of the median survival line.}

\item{pval_pos}{a vector of 2 numeric, coordinate positions of the p-value text (\code{c(x,y)}). Default in \code{NULL}, should appear on the bottom left.}

\item{scale_x_breaks}{passed to \code{ggplot2::scale_x_continuous(breaks = )} of the Kaplan-Meier curve and of the risk table. Default is \code{ggplot2::waiver()} (default setup of ggplot2). Overrules \code{scale_x_break_by}.}

\item{scale_x_break_by}{a numeric, interval width when the x-axis (i.e. time) will be broken (e.g. every \code{4} weeks), as well as the times when the number of patient at risk is calculated. Default is \code{NULL} (calls the default setup of ggplot2). Ignored if \code{scale_x_breaks} is provided.}

\item{scale_y_labels}{passed to \code{ggplot2::scale_y_continuous(labels = )} when the Kaplan-Meier curve is constructed. Default is \code{scales::percent} to label the survival as percentage. Set to \code{identity} or \code{ggplot2::waiver()} for the original numeric values.}

\item{scale_y_risktable_reverse}{a logical, should the y-axis of the risk table be reversed? By default, the levels of a factor variable are printed from bottom to top, but the human eyes expect to read a table from top to bottom.}

\item{label_x}{a character, the name of the x-axis (i.e. time) of the Kaplan-Meier figure and of the risk table.}

\item{label_y_fig}{a character, the name of the y-axis (i.e. survival) of the Kaplan-Meier figure.}

\item{label_color}{a character, the name of the legend for colors (and, by default, of the y-axis of the risk table). Can also be an expression (\code{expression()}) if special character or super/subscripts should be shown.}

\item{label_y_risktab}{a character, the name of the y-axis of the risk table. Default is the same as the name of the legend for colors \code{label_color}.}

\item{legend_label_kpm}{a character, the label of the legend that describes the line type of the Kaplan-Meier curve.}

\item{legend_label_censor}{a character, the label of the legend that describes the shape of the censoring events.}

\item{legend_label_se}{a character, the label of the legend that describes the ribbon of parametric confidence interval. Default is computed as function of \code{ci_level}.}

\item{legend_label_median}{a character, the label of the legend that describes the line type of the median survival.}

\item{legend_label_sim}{a character, the label of the legend that describes the ribbon of simulation-based confidence interval. Default is computed as function of \code{ci_level}.}

\item{guide_legend_nrow}{a numeric, the number of rows of the color legend. Set a high number if you have many color categories.}

\item{scale_color_labels}{passed to \code{ggplot2::scale_color_manual(labels = )} for the Kaplan-Meier figure, and \code{ggplot2::scale_y_discrete(labels = )} for the risk table. The default is \code{waiver()} (calls the default setup of ggplot2), but can accept \code{scales::label_parsed} if the coloring variable is an expression (\code{expression()}) and special character or super/subscripts should be shown.}

\item{scale_color_values}{passed to \code{ggplot2::scale_color_manual(values = )} for the Kaplan-Meier figure and the risk table. Default will use the default PMX graphic charter, but named vectors of colors (with names matching with factor levels) are welcome. Levels not available in the data will be dropped.}

\item{facetting_args}{a list, arguments that will overwrite the default arguments of \code{facet_wrap()}. Note that it will not overwrite \code{vars}, the latter being informed by \code{facet_var}. Use \code{list(labeller = ggplot2::label_parsed)} if the facetting variable is an expression (\code{expression()}) and special character or super/subscripts should be shown.}

\item{xlim}{a vector of 2 numeric values, limits of the x-axes of the Kaplan-Meier plot and risk table. Default will use the default ggplot setup. Passed to \code{ggplot2::coord_cartesian()}.}

\item{ylim}{a vector of 2 numeric values, limits of the y-axis of the Kaplan-Meier figure. Default is 0 to 1. Passed to \code{ggplot2::coord_cartesian()}.}

\item{title_risktable}{a character, the title above the risk table.}

\item{arrange}{a logical, should the elements of the plot be arranged in a single plot with \code{cowplot::grid.arrange()} (default is \code{TRUE}). If \code{FALSE}, a list with separated elements will be returned (useful for advanced customization.)}

\item{rel_heights}{a list specifying the relative heights (1) between the legends (default is 1 and 1, same height), (2) between the Kaplan-Meier figure and the risk table (default is 2 and 1, i.e. figure twice higher than table) and (3) overall (default is 1 and 8, combined legends takes 1/9 of heights, combined figure and tables takes 8/9 of heights)}

\item{check_input}{a logical, should the data format be checked?}
}
\value{
by default, a single "ggplot" object arranged with \code{cowplot::plot.grid()}. If \code{arrange = FALSE}, a list of ggplot objects.
}
\description{
Single function for graphical exploration of survival data,
model evaluation of time-to-event model (VPC) and visualization of simulated data.
If only observed data (\code{data}) is supplied, the Kaplan-Meier curve, censoring events, median survival time,
risk table and parametric confidence interval are
automatically reported, as well as log rank test p value if data is stratified by coloring groups (\code{color_var}).
If both observed and simulated data (\code{sdata}) are supplied, censoring events, risk table and
simulation-based confidence interval are reported.
If only simulated data is supplied, the simulation-based prediction interval is reported.
Choose to show/hide any of these elements with the \verb{show_*} arguments.
}
\details{
Data can be grouped/colored with the \code{color_var} argument, and faceted with the \code{facet_var} argument. Preferably, these should refer to factor variables.

Arrange the relative heights between the legends, the figure and the risk table with the \code{rel_heights} argument.
}
\examples{

dat <- PMXtte::simplettedata
simdat <- PMXtte::simplettesimdata

ggKAP(dat) # Graphical exploration
ggKAP(dat, simdat) # VPC
ggKAP(sdata = simdat) # Simulation only

ggKAP(dat, show_kpm = FALSE)
ggKAP(dat, show_censor = FALSE)
ggKAP(dat, show_se = FALSE)
ggKAP(dat, show_risktable = FALSE)
ggKAP(dat, show_median = FALSE)
try(ggKAP(dat, show_sim = TRUE), TRUE)

ggKAP(dat, cuminc = TRUE)

ggKAP(dat, color_var = "SEXF")
ggKAP(dat, color_var = "SEXF", label_color = "Sex")
ggKAP(dat, color_var = "SEXF", show_pval = FALSE)
ggKAP(dat, color_var = "SEXF", scale_color_values = c("red", "yellow"))
ggKAP(dat, color_var = "SEXF", scale_color_values = c(Male = "yellow", Female = "red"))
ggKAP(dat, color_var = "SEXF", ci_alpha = .3)
ggKAP(dat, color_var = "SEXF", ci_alpha = .7)

ggKAP(dat, color_var = "SEXF", ci_level = .5)
ggKAP(
  dat, ci_alpha = .2, censor_shape = 19,
  censor_size = 6, median_linetype = 3
)

ggKAP(dat, scale_x_breaks = c(0, 30, 40, 80, 100)) #Both risk table and graph are updated
ggKAP(dat, scale_x_break_by = 12) #Or use this shortcut if you just want to play with the width of interval

ggKAP(
  dat, color_var = "SEXF",
  xlim = c(0, 50), ylim = c(.10, .5)
) # limits do not change the underlying data (e.g.: same p value)

ggKAP(
  dat, color_var = "SEXF",
  pval_pos = c(80, .8)
) # manually define p value position on the top right

# Labelling

ggKAP(dat, color_var = "AUCQF")
ggKAP(dat, color_var = "AUCQF", guide_legend_nrow = 2) # force the number of rows

# This will update both the "color legend" label and the y axis of the risk table
ggKAP(dat, color_var = "AUCQF", label_color = "Nice color label")

# But it can be differentiated if needed
ggKAP(dat, color_var = "AUCQF",
  label_color = "Nice color label",
  label_y_risktab = "Label for the\ny-axis of risk table"
  )



ggKAP(dat, rel_heights = list(
  legends = c(1,1),
  figtable = c(8,2),
  overall = c(1,8)
))
ggKAP(dat, color_var = "SEXF", facet_var = "AUCQF", rel_heights = list(
  legends = c(1,1),
  figtable = c(8,4),
  overall = c(2,8)
))

ggKAP(dat, simdat, facet_var = "SEXF")
ggKAP(sdata = simdat, facet_var = "SEXF")

# Special characters

ggKAP(dat, color_var = "AUCQF")
ggKAP(dat, color_var = "AUCQF", scale_y_risktable_reverse = TRUE) # 1st quartile on first row
ggKAP(dat, color_var = "AUCQF", label_color = expression(AUC['0-24h']))
dat$AUCQF2 <- factor(dat$AUCQF, labels = c("1^st", "2^nd~quartile", "The~3^rd~q"))
ggKAP(dat, color_var = "AUCQF2", label_color = expression(AUC['0-24h']), scale_color_labels = scales::label_parse())

ggKAP(dat, facet_var = "AUCQF2")
ggKAP(dat, facet_var = "AUCQF2", facetting_args = list(labeller = ggplot2::label_parsed, nrow = 2))
}
