% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmmc.R
\name{ggKMMC}
\alias{ggKMMC}
\title{KMMC plots for TTE and RTTE models}
\usage{
ggKMMC(
  odata,
  sdata,
  cov,
  timevar = TIME,
  bins,
  xlab = waiver(),
  ylab = "Mean of covariate",
  obsCol = "black",
  fill = PMXtte:::PMXColors_pmx_palettes(name = "light"),
  CI = 90,
  ci = NULL,
  ...
)
}
\arguments{
\item{odata}{the observed data with one row per subject with a time and event indicator (1==event, 0==censored)}

\item{sdata}{the simulated data with one row per subject and replicate with a time and event indicator (1==event, 0==censored)}

\item{timevar}{The x-axis variable. Default is TIME.}

\item{bins}{The bins of the timevar to calculate the y-variable for.}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{obsCol}{color for the observed line}

\item{fill}{The fill color for the simulation interval.}

\item{ci}{what level of confidence to we want for the simulated Kaplan-Meier}

\item{...}{}
}
\description{
KMMC plots for TTE and RTTE models
}
\examples{
rttedata <- readr::read_csv(system.file('extdata/DAT-1c-RED-1a-PMX-WOWTTE-PFPMX-1.csv', package= 'PMXtte'), show_col_types = FALSE)
rttedata <- dplyr::filter(rttedata, EVID == 0, TYPE == 2)

# Load observed data
odata <- rttedata \%>\%
  dplyr::filter(STUDYIDN == 1) \%>\%
  dplyr::mutate(
    TIME = TSFDW,
    SEXNF = factor(SEXN, levels = c(1,2), c("Male", "Female"))
  )

# Load simulated data
sdata <- readRDS(system.file('extdata/vpcdat.rds', package= 'PMXtte'))
sdata <- sdata \%>\%
  dplyr::mutate(
    TIME = TIME / (24*7)
  ) \%>\%
  dplyr::group_by(ITER, ID) \%>\%
  dplyr::mutate(EVCOUNT = cumsum(DV)) \%>\% # Compute EVCOUNT
  dplyr::ungroup()

odata1 <- odata \%>\% filter_xth_event(1) \%>\% mutate(TIME2 = TIME)
sdata1 <- sdata \%>\% filter_xth_event(1) \%>\%
  mutate(TIME2 = TIME) \%>\%
  rename(DOSEN=DOSE)

ggKMMC(odata1,
       sdata1,
       cov = DOSEN,
       bins=seq(from=0,to=52,by=1)) +
  theme(legend.position="top") +
  theme(legend.title = element_blank())

}
