% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummaryTTE.R
\name{makeSummaryTableTTE}
\alias{makeSummaryTableTTE}
\title{Title}
\usage{
makeSummaryTableTTE(
  df,
  myID = "ID",
  myDV = "DV",
  digits = 3,
  nIdColNm = "\\\\textbf{nID\\\\textsuperscript{a}}",
  nEventColNm = "\\\\textbf{nEvent\\\\textsuperscript{b}}",
  avnEventColNm = "\\\\textbf{pEvent/nID\\\\textsuperscript{c}}",
  caption = "Number of patients and number of events",
  label = "tab:anaSummary",
  footnote =
    "\\\\textsuperscript{a}Number of patients\\\\newline\\\\textsuperscript{b}Number of events\\\\newline\\\\textsuperscript{c}Proportion of number of events",
  ...
)
}
\arguments{
\item{nEventColNm}{the charactrer string to be printed as the column name with number of events}

\item{avnEventColNm}{the charactrer string to be printed as the column name with average number of events}
}
\description{
Title
}
