% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummaryTTE.R
\name{makeSummaryTableTTE}
\alias{makeSummaryTableTTE}
\title{Single data summary table for (R)TTE data}
\usage{
makeSummaryTableTTE(
  df,
  myID = "ID",
  myDV = "DV",
  digits = 3,
  showAvnObs = !isRTTE(df, myID = myID),
  nObsColNm = "\\\\textbf{nEvent\\\\textsuperscript{b}}",
  avnObsColNm = "\\\\textbf{pEvent\\\\textsuperscript{c}}",
  caption = "Number of patients and number of events",
  footnote = paste0("\\\\textsuperscript{a}Number of patients\\\\newline",
    "\\\\textsuperscript{b}Number of events\\\\newline", if (showAvnObs) {
    
    "\\\\textsuperscript{c}Proportion of number of events"
 }),
  myFun = NULL,
  ...
)
}
\arguments{
\item{df}{is the data frame}

\item{myID}{is the name of the ID column, default is "ID"}

\item{myDV}{is the name of the DV column, default is "DV"}

\item{digits}{is the number of significant digits}

\item{showAvnObs}{a logical, show a column with the average number of events? Default will be \code{FALSE} if the data is recognized as RTTE data, \code{TRUE} if not.}

\item{nObsColNm}{is the character string to be printed as the column name with the number of observations, default is "\\\\textbf\{nEvent\\\\textsuperscript\{b\}\}"}

\item{avnObsColNm}{is the character string to be printed as the column name with the average number of observations per subject in a given strata, default is "\\\\textbf\{pEvent\/nID\\\\textsuperscript\{c\}\}"}

\item{caption}{is the table caption. Assign NULL to this argument produce table without caption. Default is "Number of patients and number of events"}

\item{footnote}{is the text for footnote, default is "\\\\textsuperscript\{a\}Number of patients\\\\newline\\\\textsuperscript\{b\}Number of events\\\\newline\\\\textsuperscript\{c\}Proportion of number of events" for TTE data.}

\item{myFun}{internal function for the calculation of summarized data. If NULL, the default, an internal function specific to R(TTE) data is used. This should not be changed for a standard use.}

\item{...}{passed to \code{PhRame_makeSummaryTable()}, possibly for the additional optional arguments compatible with latex}
}
\value{
By default, the table at Latex format. Alternatively, if \code{asList = TRUE}, a list of data.frame tables
}
\description{
Derive a single data summary table with up to 2 levels of stratification, specifically tailored for (R)TTE data.
It is a wrapper around \code{PhRame_makeSummaryTable()}, which means that it is the same function but defaulted with arguments
that use "event" instead of "observation". Any argument available in the documentation of \code{PhRame_makeSummaryTable()} can
be used (except for \code{myFun}, see the details in \code{Arguments} below). By default, the average number of event will be displayed
only for single event data, not for RTTE data.
}
\examples{
rttedata <- readr::read_csv(system.file('extdata/DAT-1c-RED-1a-PMX-WOWTTE-PFPMX-1.csv', package= 'PMXtte'), show_col_types = FALSE)
rttedata <- dplyr::filter(rttedata, EVID == 0, TYPE == 2)
# create summary as a list
makeSummaryTableTTE(rttedata,
                    outerLevel   ="STUDYIDN" ,
                    outerLabel   = "Study",
                    innerLevel   = "DOSEN",
                    innerLabel   = "Dose",
                    asList = TRUE)

# output as latex table
makeSummaryTableTTE(rttedata,
                    outerLevel   ="STUDYIDN" ,
                    outerLabel   = "Study",
                    innerLevel   = "DOSEN",
                    innerLabel   = "Dose")

# Time to first event only
tte1data <- filter_xth_event(rttedata, event = 1)
makeSummaryTableTTE(tte1data,
                    outerLevel   ="STUDYIDN" ,
                    outerLabel   = "Study",
                    innerLevel   = "DOSEN",
                    innerLabel   = "Dose",
                    asList = TRUE)
}
