% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryCountRTTE.R
\name{summaryCountRTTE}
\alias{summaryCountRTTE}
\title{Single data summary table for RTTE data}
\usage{
summaryCountRTTE(
  df,
  addFactors = FALSE,
  filterExpr = NULL,
  myID = "ID",
  myDV = "DV",
  myEVCOUNT = "EVCOUNT",
  outerLevel = NULL,
  innerLevel = NULL,
  outerLabel = NULL,
  innerLabel = NULL,
  digits = 3,
  lumpCount = Inf,
  dropCount0 = FALSE,
  nIdColNm = "\\\\textbf{nID\\\\textsuperscript{a}}",
  caption = "Number of repeated events",
  label = "tab:anaSummary",
  footnote = "\\\\textsuperscript{a}Number of subjects",
  footnoteSize = "footnotesize",
  textSize = "small",
  here = TRUE,
  numeric.dollar = FALSE,
  asList = FALSE,
  saveTex = FALSE,
  fileName = NULL,
  filePath = "./",
  myFun = NULL,
  ...
)
}
\arguments{
\item{df}{is the data frame (e.g. anaDataF) or data object created using
\code{pmxdata::derived_data} (e.g. DATobject)}

\item{addFactors}{if \strong{TRUE}, factors are added to the input data.
Default is \strong{FALSE}. This option is used iff \code{df} is a data
object}

\item{filterExpr}{is an expression used to filter the data. Either a formula
using non standard evaluation (e.g. \code{quo(AMT==0 & TYPE==1)}) or a
character string with filter expression (e.g. "(AMT==0 & TYPE==1)") is
acceptable to this argument.}

\item{myID}{is the name of the ID column, default is \strong{"ID"}}

\item{myDV}{is the name of the DV column, default is \strong{"DV"}}

\item{myEVCOUNT}{is the name of the EVCOUNT column, default is \strong{"EVCOUNT"}}

\item{outerLevel}{is the outler level (1st level) of stratification variable,
eg. "STUDYF"}

\item{innerLevel}{is the inner level (2nd level) of stratification variable,
eg. "DOSEF"}

\item{outerLabel}{is the label for the outler level (1st level) of
stratification variable, eg. "Study")}

\item{innerLabel}{is the label for the inner level (2nd level) of
stratification variable, eg. "Dose"}

\item{digits}{is the number of significant digits}

\item{lumpCount}{is the maximum number of event to display, e.g. if \code{lumpcount=3},
the category "3 or more" will be display. Only works for Latex code.
Default is \strong{Inf} (no lumping)}

\item{dropCount0}{is a logical, should categories with zero event be dropped.
Only works for Latex code. Default is \strong{FALSE} (don't drop)}

\item{nIdColNm}{is the character string to be printed as the column name
with the number of subjects, default is
\strong{"\\\\textbf\{nID\\\\textsuperscript\{a\}\}"}}

\item{caption}{is the table caption. Assign \strong{NULL} to this argument
produce table without caption. Default is \strong{"PK analysis data set:
Number of subjects with observations and number of observations"}}

\item{label}{is the label for the table used for cross-reference, default is
\strong{"tab:anaSummary"}}

\item{footnote}{is the text for footnote, default is
\strong{"\\textsuperscript\{a\} Number of
subjects\\\\newline\\\\textsuperscript\{b\} Number of
observations\\\\newline\\\\textsuperscript\{c\} Average number of
observations per subject"}}

\item{footnoteSize}{is the footnote text size, default is
\strong{"footnotesize"}}

\item{textSize}{is the text size, default is \strong{"small"}}

\item{here}{is the logical argument determining the placement of the LaTex
table. If set to \strong{TRUE} with optional argument \code{table.env=TRUE}
with \code{longtable=FALSE}, will cause the LaTeX table environment to be
set up with option H to guarantee that the table will appear exactly where
you think it will in the text. Default is \strong{TRUE}}

\item{numeric.dollar}{is the logical argument determining if the numbers will
be placed within $ symbols, default is \strong{FALSE}}

\item{asList}{if \strong{TRUE} the data summary is returned as a list where
different elements in the list contain summary of different stratification
levels. If \strong{FALSE} the data summary is returned as a latex table. If
no stratification variable is provided, this function always returns a list
of two items (Number of subjects, Number of observarions and Average number
of observarions per subject)}

\item{saveTex}{is the logical argument determining if the LaTex code for the
table will be saved on disk. Default is \strong{FALSE}}

\item{fileName}{is the name of the Tex file (without extension) with the
LaTex code for the table saved on disk, eg. if \code{fileName="mytable"} is
provided, a file called \code{"mytable.tex"} will be saved on disk. Default
is \strong{NULL}.}

\item{filePath}{is the path to the directory where the Tex file is saved.
Default is \strong{"/"}.}

\item{myFun}{internal function for the calculation of summarized data. If
NULL, the default, an internal function specific to R(TTE) data is used.
This should not be changed for a standard use.}

\item{...}{is the additional optional arguments compatible with latex}
}
\value{
By default, the table at Latex format. Alternatively, if
\code{asList = TRUE}, a list of data.frame tables.
}
\description{
Derive a single data summary table with up to 2 levels of stratification, counting the number of events in each group.
It is similar to \code{PhRame_makeSummaryTable()}, the same arguments should apply the same to both functions. There is a possibility to
lump counts of event (e.g., if \code{lumpCount = 3}, a column "3 or more" will be shown). It is possible to remove
counts if there are no events with \code{dropCount0 = TRUE}.
}
\examples{
rttedata <- readr::read_csv(system.file('extdata/DAT-1c-RED-1a-PMX-WOWTTE-PFPMX-1.csv', package= 'PMXtte'), show_col_types = FALSE)
rttedata <- dplyr::filter(rttedata, EVID == 0, TYPE == 2)
# create summary as a list
summaryCountRTTE(rttedata,
                    outerLevel   ="STUDYIDN" ,
                    outerLabel   = "Study",
                    innerLevel   = "DOSEN",
                    innerLabel   = "Dose",
                    asList = TRUE)


# output as latex table
summaryCountRTTE(rttedata,
                    outerLevel   ="STUDYIDN" ,
                    outerLabel   = "Study",
                    innerLevel   = "DOSEN",
                    innerLabel   = "Dose")
summaryCountRTTE(rttedata,
                    outerLevel   ="STUDYIDN" ,
                    outerLabel   = "Study",
                    innerLevel   = "DOSEN",
                    innerLabel   = "Dose",
                    lumpCount    = 3
                    )
}
