% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFollowUpTime.R
\name{summaryFollowUpTime}
\alias{summaryFollowUpTime}
\title{Title}
\usage{
summaryFollowUpTime(
  df,
  myID = "ID",
  myDV = "DV",
  time = "TSFDH",
  timeConversion = 1,
  count = "EVCOUNT",
  outerLevel = NULL,
  innerLevel = NULL,
  outerLabel = NULL,
  innerLabel = NULL,
  digits = 3,
  digits_rate = 2,
  nEventColNm = "\\\\textbf{nEvent\\\\textsuperscript{a}}",
  fTimeColNm = "\\\\textbf{TotalFollowUpTime\\\\textsuperscript{b}}",
  EventRateColNm = "\\\\textbf{AnnualizedEventRate/nID\\\\textsuperscript{c}}",
  caption = "Annual Follow Up time",
  label = "tab:anaSummary",
  footnote =
    "\\\\textsuperscript{a}Number of events\\\\newline\\\\textsuperscript{b}Sum of individual follow-up times per group.\\\\newline\\\\textsuperscript{c}Annual event rate",
  asList = FALSE,
  ...
)
}
\arguments{
\item{time}{Column used to calculate annual time}

\item{timeConversion}{factor to multiply the time column by to convert it to another format(e.g. annual)}

\item{count}{Column that contains count for events}

\item{digits}{the number of significant digits for the follow up time column}

\item{digits_rate}{the number of significant digits for the Event rate column}

\item{nEventColNm}{the charactrer string to be printed as the column name with the number of events}

\item{fTimeColNm}{the charactrer string to be printed as the column name with follow up time}

\item{EventRateColNm}{the charactrer string to be printed as the column name with Event rate}
}
\description{
Title
}
\examples{
ttedata  <- readr::read_csv(system.file('extdata/anaDATAf.csv', package= 'PMXtte'))

#create summary as a list
summaryFollowUpTime(ttedata,
                    outerLevel   = "DOSEN",
                    outerLabel   = "Dose",
                    innerLevel   = "STUDYIDN",
                    innerLabel   = "Study",
                    asList = TRUE,
                    timeConversion = 1/24/365.25)

#output as latex table
summaryFollowUpTime(ttedata,
                    outerLevel   = "DOSEN",
                    outerLabel   = "Dose",
                    innerLevel   = "STUDYIDN",
                    innerLabel   = "Study",
                    timeConversion = 1/24/365.25)
}
