% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFollowUpTime.R
\name{summaryFollowUpTime}
\alias{summaryFollowUpTime}
\title{Single data summary table for follow-up times with RTTE data}
\usage{
summaryFollowUpTime(
  df,
  myID = "ID",
  myDV = "DV",
  myTIME = "TSFDH",
  timeConversion = 1/(24 * 365.25),
  myEVCOUNT = "EVCOUNT",
  outerLevel = NULL,
  innerLevel = NULL,
  outerLabel = NULL,
  innerLabel = NULL,
  digits = 0,
  digits_rate = 2,
  nEventColNm = "\\\\textbf{nEvent\\\\textsuperscript{a}}",
  fTimeColNm = "\\\\textbf{Observation period (year)\\\\textsuperscript{b}}",
  EventRateColNm = "\\\\textbf{Event rate (1/year)\\\\textsuperscript{c}}",
  caption =
    "Summary of observation periods and annual event rates in the analysis data set.",
  label = "tab:anaSummaryFollowupTime",
  footnote =
    "\\\\textsuperscript{a}Total number of events\\\\newline\\\\textsuperscript{b}\\n                                 The observation period (total) is the sum of the individual observation periods per group.\\n                                 \\\\newline\\\\textsuperscript{c}Annual event rate is calculated, per group, as the total number of events divided by the observation period (total).",
  asList = FALSE,
  ...
)
}
\arguments{
\item{df}{is the data frame (e.g. anaDataF) or data object created using
\code{pmxdata::derived_data} (e.g. DATobject)}

\item{myID}{is the name of the ID column, default is \strong{"ID"}}

\item{myDV}{is the name of the DV column, default is \strong{"DV"}}

\item{myTIME}{Column used to calculate annual time, default is \code{"TSFDH"}}

\item{timeConversion}{factor to multiply the time column by to convert it to another format (e.g. annual). Default is \code{1/(24*365.25)} because default myTIME is \code{"TSFDH"}}

\item{myEVCOUNT}{Column that contains count for events, default is \code{"EVCOUNT"}}

\item{outerLevel}{is the outer level (1st level) of stratification variable,
eg. "STUDYNF"}

\item{innerLevel}{is the inner level (2nd level) of stratification variable,
eg. "DOSENF"}

\item{outerLabel}{is the label for the outer level (1st level) of
stratification variable, eg. "Study")}

\item{innerLabel}{is the label for the inner level (2nd level) of
stratification variable, eg. "Dose"}

\item{digits}{the number of significant digits for the follow up time column}

\item{digits_rate}{the number of significant digits for the Event rate column}

\item{nEventColNm}{the character string to be printed as the column name with the number of events}

\item{fTimeColNm}{the character string to be printed as the column name with follow up time}

\item{EventRateColNm}{the character string to be printed as the column name with Event rate}

\item{caption}{is the table caption. Assign NULL to this argument produce table without caption.}

\item{label}{is the label for the table used for cross-reference, default is
\strong{"tab:anaSummary"}}

\item{footnote}{is the text for footnote, default is "\\textsuperscript{a}Number of events\\newline\\textsuperscript{b}Total
observation period calculated as sum of individual observation period per row.
\\newline\\textsuperscript{c}Annual event rate calculated, for each row, as number of events per
group divided by total observation period per group."#' @param label is the label for the table used for cross-reference.}

\item{asList}{if \strong{TRUE} the data summary is returned as a list where
different elements in the list contain summary of different stratification
levels. If \strong{FALSE} the data summary is returned as a latex table. If
no stratification variable is provided, this function always returns a list
of two items (Number of event, Observation period, Event rate)}

\item{...}{is the additional optional arguments compatible with
\link[Hmisc:latex]{Hmisc::latex}}
}
\value{
By default, the table at Latex format. Alternatively, if \code{asList = TRUE}, a list of data.frame tables
}
\description{
Derive a single data summary table with up to 2 levels of stratification, in order to display the number
of events, the total follow-up time and the annualized event rates. It is a wrapper around \code{PhRame_makeSummaryTable()},
which means that it is the same function but defaulted with arguments that use "event" instead of "observation".
Any argument available in the documentation of \code{PhRame_makeSummaryTable()} can be used (except for \code{myFun},
see the details in \code{Arguments} below). New arguments include the possibility to handle the Time and Event Counts
variables, and possibly conversion in the relevant units.
}
\examples{
rttedata <- readr::read_csv(system.file('extdata/DAT-1c-RED-1a-PMX-WOWTTE-PFPMX-1.csv', package= 'PMXtte'), show_col_types = FALSE)
rttedata <- dplyr::filter(rttedata, EVID == 0, TYPE == 2)
rttedata$TRTF <- factor(paste(rttedata$TRTN, "mg"), levels = paste(sort(unique(rttedata$TRTN)), "mg"))

#create summary as a list
summaryFollowUpTime(rttedata,
                    outerLevel   ="STUDYIDN" ,
                    outerLabel   = "Study",
                    innerLevel   = "TRTF",
                    innerLabel   = "Treatment",
                    asList = TRUE,
                    timeConversion = 1/24/365.25)

#output as latex table
summaryFollowUpTime(rttedata,
                    outerLevel   ="STUDYIDN" ,
                    outerLabel   = "Study",
                    innerLevel   = "TRTF",
                    innerLabel   = "Treatment",
                    timeConversion = 1/24/365.25)
}
